Directory and db schemas refactor
