+++ app/routes/dashboard.new/route.tsx
@@ -1,12 +1,7 @@
-import {
-    ActionFunctionArgs,
-    LoaderFunctionArgs,
-    json,
-    redirect,
-} from "@remix-run/node";
-import { Form, NavLink, useActionData, useLoaderData } from "@remix-run/react";
+import { ActionFunctionArgs, json, redirect } from "@remix-run/node";
+import { Form, NavLink, useActionData } from "@remix-run/react";
 import { validate } from "./validate";
-import { getEmpleado, updateEmpleado } from "./queries";
+import { createEmpleado } from "./queries";
 
 export async function action({ request }: ActionFunctionArgs) {
     const formData = await request.formData();
@@ -22,10 +17,7 @@ export async function action({ request }: ActionFunctionArgs) {
     const studies = String(formData.get("studies"));
     const cond = String(formData.get("cond"));
     const area = String(formData.get("area"));
-    const a_cargo = String(formData.get("a_cargo"));
     const disp = String(formData.get("disp"));
-    const url = new URL(request.url);
-    const id = url.pathname.split("/").pop();
 
     const errors = await validate(
         name,
@@ -39,7 +31,6 @@ export async function action({ request }: ActionFunctionArgs) {
         studies,
         cond,
         area,
-        a_cargo,
         disp
     );
 
@@ -47,8 +38,7 @@ export async function action({ request }: ActionFunctionArgs) {
         return json({ ok: false, errors });
     }
 
-    const empleado = await updateEmpleado({
-        id,
+    const empleado = await createEmpleado({
         name,
         DNI,
         birth,
@@ -60,7 +50,6 @@ export async function action({ request }: ActionFunctionArgs) {
         studies,
         cond,
         area,
-        a_cargo,
         disp,
     });
 
@@ -68,7 +57,7 @@ export async function action({ request }: ActionFunctionArgs) {
         return json({ ok: false, errors: { a_cargo: "Datos inválidos" } });
     }
 
-    return redirect("/dashboard/nomina");
+    return redirect("/dashboard");
 }
 
 interface ActionData {
@@ -83,50 +72,12 @@ interface ActionData {
         cond?: string;
         studies?: string;
         area?: string;
-        a_cargo?: string;
         disp?: string;
     } | null;
 }
 
-interface LoaderData {
-    empleado: {
-        name: string;
-        DNI: string;
-        birth: Date;
-        kids: number;
-        address: string;
-        tel: string;
-        obvs: string;
-        workData: {
-            ant: string;
-            cond: string;
-            studies: string;
-            area: string;
-            a_cargo: string;
-            disp: string;
-        };
-    };
-}
-
-export async function loader({ request }: LoaderFunctionArgs) {
-    const url = new URL(request.url);
-    const id = url.pathname.split("/").pop();
-    const empleado = await getEmpleado(String(id));
-    return { empleado };
-}
-
-function formateDate(dateString: Date) {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    return `${year}-${month}-${day}`;
-}
-
-export default function EmpleadoPage() {
+export default function NuevoEmpleadoPage() {
     const actionData: ActionData = useActionData<typeof action>();
-    const loaderData: LoaderData = useLoaderData<typeof loader>();
-    const empleado = loaderData.empleado;
 
     const nameError = actionData?.errors?.name;
     const DNIError = actionData?.errors?.DNI;
@@ -138,7 +89,6 @@ export default function EmpleadoPage() {
     const studiesError = actionData?.errors?.studies;
     const condError = actionData?.errors?.cond;
     const areaError = actionData?.errors?.area;
-    const a_cargoError = actionData?.errors?.a_cargo;
     const dispError = actionData?.errors?.disp;
 
     return (
@@ -160,12 +110,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="name"
-                            id="name"
-                            defaultValue={empleado.name}
-                        />
+                        <input type="text" name="name" id="name" />
 
                         <label htmlFor="DNI">
                             DNI{" "}
@@ -173,12 +118,7 @@ export default function EmpleadoPage() {
                                 <span className="text-red-600">{DNIError}</span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="DNI"
-                            id="DNI"
-                            defaultValue={empleado.DNI}
-                        />
+                        <input type="text" name="DNI" id="DNI" />
 
                         <label htmlFor="birth">
                             Fecha de Nacimiento{" "}
@@ -188,12 +128,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="date"
-                            name="birth"
-                            id="birth"
-                            defaultValue={formateDate(empleado.birth)}
-                        />
+                        <input type="date" name="birth" id="birth" />
 
                         <label htmlFor="kids">
                             Cantidad de Hijos{" "}
@@ -203,12 +138,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="number"
-                            name="kids"
-                            id="kids"
-                            defaultValue={empleado.kids}
-                        />
+                        <input type="number" name="kids" id="kids" />
 
                         <label htmlFor="tel">
                             Teléfono{" "}
@@ -216,12 +146,7 @@ export default function EmpleadoPage() {
                                 <span className="text-red-600">{telError}</span>
                             )}
                         </label>
-                        <input
-                            type="tel"
-                            name="tel"
-                            id="tel"
-                            defaultValue={empleado.tel}
-                        />
+                        <input type="tel" name="tel" id="tel" />
 
                         <label htmlFor="address">
                             Domicilio{" "}
@@ -231,12 +156,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="address"
-                            id="address"
-                            defaultValue={empleado.address}
-                        />
+                        <input type="text" name="address" id="address" />
 
                         <label htmlFor="obvs">Obvservaciones </label>
                         <textarea
@@ -245,10 +165,8 @@ export default function EmpleadoPage() {
                             cols={30}
                             rows={5}
                             style={{ resize: "none" }}
-                            defaultValue={empleado.obvs}
                         ></textarea>
                     </div>
-
                     <div style={{ display: "flex", flexDirection: "column" }}>
                         <label htmlFor="ant">
                             Antigüedad{" "}
@@ -256,12 +174,7 @@ export default function EmpleadoPage() {
                                 <span className="text-red-600">{antError}</span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="ant"
-                            id="ant"
-                            defaultValue={empleado.workData.ant}
-                        />
+                        <input type="text" name="ant" id="ant" />
 
                         <label htmlFor="studies">
                             Estudios{" "}
@@ -271,12 +184,29 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="studies"
-                            id="studies"
-                            defaultValue={empleado.workData.studies}
-                        />
+                        <select name="studies" id="studies">
+                            <option value="Primario Completo">
+                                Primario Completo
+                            </option>
+                            <option value="Secundario Incompleto">
+                                Secundario Incompleto
+                            </option>
+                            <option value="Secundario Incompleto">
+                                Secundario Incompleto
+                            </option>
+                            <option value="Terciario Incompleto">
+                                Terciario Incompleto
+                            </option>
+                            <option value="Terciario Completo">
+                                Terciario Completo
+                            </option>
+                            <option value="Universitario Incompleto">
+                                Universitario Incompleto
+                            </option>
+                            <option value="Universitario Completo">
+                                Universitario Completo
+                            </option>
+                        </select>
 
                         <label htmlFor="cond">
                             Condición{" "}
@@ -286,11 +216,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <select
-                            name="cond"
-                            id="cond"
-                            defaultValue={empleado.workData.cond}
-                        >
+                        <select name="cond" id="cond">
                             <option value="Contrato">Contrato</option>
                             <option value="Planta Provincial">
                                 Planta Provincial
@@ -311,27 +237,29 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="area"
-                            id="area"
-                            defaultValue={empleado.workData.area}
-                        />
-
-                        <label htmlFor="a_cargo">
-                            Persona a cargo{" "}
-                            {a_cargoError && (
-                                <span className="text-red-600">
-                                    {a_cargoError}
-                                </span>
-                            )}
-                        </label>
-                        <input
-                            type="text"
-                            name="a_cargo"
-                            id="a_cargo"
-                            defaultValue={empleado.workData.a_cargo}
-                        />
+                        <select name="area" id="area">
+                            <option value="RR.HH">RR.HH</option>
+                            <option value="Administración">
+                                Administración
+                            </option>
+                            <option value="Formadores Deportivos">
+                                Formadores Deportivos
+                            </option>
+                            <option value="Comunicación">Comunicación</option>
+                            <option value="Juventud">Juventud</option>
+                            <option value="Tareas Operativas">
+                                Tareas Operativas
+                            </option>
+                            <option value="Gestión Ciudadana">
+                                Gestión Ciudadana
+                            </option>
+                            <option value="Promoción Institucional">
+                                Promoción Institucional
+                            </option>
+                            <option value="Tareas Generales">
+                                Tareas Generales
+                            </option>
+                        </select>
 
                         <label htmlFor="disp">
                             Disponibilidad horaria{" "}
@@ -341,12 +269,7 @@ export default function EmpleadoPage() {
                                 </span>
                             )}
                         </label>
-                        <input
-                            type="text"
-                            name="disp"
-                            id="disp"
-                            defaultValue={empleado.workData.disp}
-                        />
+                        <input type="text" name="disp" id="disp" />
                     </div>
                 </div>
                 <div className="flex gap-8">
@@ -354,7 +277,7 @@ export default function EmpleadoPage() {
                         type="submit"
                         className="bg-slate-600 p-2 rounded-lg text-white"
                     >
-                        Actualizar
+                        Crear
                     </button>
                     <NavLink
                         to={"/dashboard/nomina"}
